requirements.txt
streamlit==1.35.0
openai>=1.0.0
pytesseract
Pillow